requirements.txt
fastapi
uvicorn
azure-functions
azure-cosmos
python-multipart
pytest
pytest-asyncio
pytest-cov
httpx
